% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allelic_analysis.R
\name{allelic_compare_p}
\alias{allelic_compare_p}
\title{Get the allelic comparison p value for the enhancers}
\usage{
allelic_compare_p(allelic_compare, enhancer)
}
\arguments{
\item{allelic_compare}{the results produced by \code{\link{allelic_compare}}}

\item{enhancer}{a list of hte enhancer variant ID}
}
\value{
a data.frame containing all the allelic comparison p value and p.adjust for the enhancers
}
\description{
Get the allelic comparison p value for the enhancers
}
