% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allelic_enhancer_dot_plot.R
\docType{methods}
\name{allelic_enhancer_dot_plot}
\alias{allelic_enhancer_dot_plot}
\title{Dot plot for significant allelic different variant}
\usage{
allelic_enhancer_dot_plot(
  data,
  log2,
  max_fold,
  label,
  foldchange = 0.25,
  type = 1,
  main = NULL
)
}
\arguments{
\item{data}{data used for plotting}

\item{log2}{column indicating the log2 fold change, character}

\item{max_fold}{column indicating maximum fold change, character}

\item{label}{column indicating the significant variant, character}

\item{foldchange}{fold change cut-off for claiming allelic enhancer variant, such as 0.25, numeric}

\item{type}{whether to include the annotation line with 1 means yes}

\item{main}{the tile of the graph}
}
\description{
Dot plot for significant allelic different variant
}
\examples{
allelic_enhancer_dot_plot(te7_ctrl_plot,"log2_aver", "max_fold", "pos")
}
