% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allgeneric.R, R/fold_enhancer_plot.R
\docType{methods}
\name{fold_enhancer_plot}
\alias{fold_enhancer_plot}
\alias{fold_enhancer_plot,DESeqResults-method}
\alias{fold_enhancer_plot.DESeqResults}
\title{Plot histogram for significant allele with all alleles as background}
\usage{
fold_enhancer_plot(object, ...)

\S4method{fold_enhancer_plot}{DESeqResults}(
  object,
  alpha = 0.05,
  bw = 0.01,
  breaks = waiver(),
  foldchange = 0.5,
  yadjust = 100,
  ladjust = 0.25,
  radjust = 1,
  xmin,
  xmax,
  ymax,
  main = deparse(substitute(object))
)
}
\arguments{
\item{object}{a \code{DESeqResults} object produced by \code{\link{results}}}

\item{...}{additional arguments}

\item{alpha}{the significance level for thresholding adjusted p-values}

\item{bw}{bandwidth}

\item{breaks}{One of:
- `NULL` for no breaks
- `waiver()` for the default breaks computed by the transformation object
- A numeric vector of positions
- A function that takes the limits as input and returns breaks as output}

\item{foldchange}{the fold change of the yellow line}

\item{yadjust}{the y location of fold change text}

\item{ladjust}{the adjustment for the left yellow line}

\item{radjust}{the adjustment for the right yellow line}

\item{xmin}{the minimum value for the x-axis}

\item{xmax}{the maximum value for the x-axis}

\item{ymax}{the maximum value for the y-axis}

\item{main}{the title of the figure.}
}
\description{
A quick plot to visualize the distribution of the significant alleles with all alleles as background. Yellow lines are added to indicate the fold change cut off.
}
\examples{
dds <- DESeqDataSetFromMatrix()
dds <- DESeq(dds)
res <- results(dds, contrast=c("condition","exp","ctrl"))
fold_enhancer_plot(res)
}
