% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allelic_analysis.R
\name{max_fold}
\alias{max_fold}
\title{Get the max fold change for each varaint}
\usage{
max_fold(object)
}
\arguments{
\item{object}{a \code{DESeqResults} object produced by \code{\link{results}}}
}
\value{
a data.frame containing variant name, max fold change and max absolute fold change
}
\description{
Get the max fold change for each varaint
}
