% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mpraprofiler.R
\docType{package}
\name{mpraprofiler}
\alias{mpraprofiler}
\title{mpraprofiler: A package of analysis and visualization for massively parallel reporter assays data}
\description{
This package provides allelic analysis and plot functions:
}
\section{Allelic Analysis}{

fold change
}

\section{Plot}{

fold_enhancer_plot
}

