% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allelic_analysis.R
\name{nr_r_ratio}
\alias{nr_r_ratio}
\title{Get the non-ref vs ref ratio for each variant}
\usage{
nr_r_ratio(cts)
}
\arguments{
\item{cts}{data counts prepared for DESeq2 analysis.}
}
\value{
a data.frame containing all the non-ref vs ref ratio with the correpsonding annotation
}
\description{
Get the non-ref vs ref ratio for each variant
}
