% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allelic_analysis.R
\name{p_value_function}
\alias{p_value_function}
\title{A quickway to get the p value for each variant, only suited for one sample t-test}
\usage{
p_value_function(x)
}
\arguments{
\item{x}{a set of data for calculating p-value}
}
\value{
p-value
}
\description{
A quickway to get the p value for each variant, only suited for one sample t-test
}
